Renaming emddetrender, adding references. Testing summation workaround.
